computational-cell/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "computational-cell"
version = "0.1.0"
description = "Precision whole-cell modeling framework for multiscale biological simulations"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["computational biology", "whole-cell modeling", "multiscale simulation", "systems biology"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "h5py>=3.8.0",
    "pyyaml>=6.0",
    "jsonschema>=4.17.0",
    "meshio>=5.3.0",
    "pandas>=2.0.0",
    "sympy>=1.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",
    "black>=23.3.0",
    "ruff>=0.0.270",
    "mypy>=1.3.0",
    "sphinx>=6.2.0",
    "sphinx-rtd-theme>=1.2.0",
    "jupyter>=1.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]
fem = [
    "fenics>=2019.1.0",  # Optional FEM backend
]
fast = [
    "numba>=0.57.0",
]
viz = [
    "pyvista>=0.39.0",
    "plotly>=5.14.0",
]
all = [
    "computational-cell[dev,fem,fast,viz]",
]

[project.urls]
Homepage = "https://github.com/yourusername/computational-cell"
Documentation = "https://computational-cell.readthedocs.io"
Repository = "https://github.com/yourusername/computational-cell.git"
Issues = "https://github.com/yourusername/computational-cell/issues"

[project.scripts]
cc-run = "cc.core.run:main"
cc-validate = "cc.validation.runner:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cc*"]

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "PT"]
ignore = ["E501"]  # Line too long (handled by black)

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --cov=cc --cov-report=term-missing --cov-report=html"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "benchmark: marks tests as benchmarks",
]

[tool.coverage.run]
source = ["cc"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
precision = 2
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
